Criação do controller Barber
